Add ApiHandler class and update Main method
A new ApiHandler class was created to manage API calls and the deserialization of responses into Person objects. The Main method in Program.cs has been refactored to asynchronously call this API handler and handle the returned list of person objects. The old code which manually serialized and deserialized a Person object has been commented out.